fix mistakes
