+++ project/src/components/favorite-location/favorite-location.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Hotel } from '../../types/hotel';
 import { PlaceCard } from '../place-card/place-card';
-import { AppRoute, cityCardType } from '../../consts/consts';
 import { Link } from 'react-router-dom';
+import { AppRoute } from '../../consts/app-route';
+import { cityCardType } from '../../consts/city-card-type';
 
 type FavoritesLocationProps = {
   city: string
