+++ project/src/components/private-route/pravate-route.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Navigate } from 'react-router-dom';
-import { AppRoute, AuthorizationStatus } from '../../consts/consts';
+import { AppRoute } from '../../consts/app-route';
+import { AuthorizationStatus } from '../../consts/authorization-status';
 
 type PrivateRouteProps = {
   authorizationStatus: AuthorizationStatus;
