+++ project/src/consts/app-route.ts
@@ -1,0 +1,6 @@
+export enum AppRoute {
+Main = '/',
+Login = '/login',
+Favorites = '/favorites',
+Room = '/offer'
+}
