+++ project/src/consts/cities.ts
@@ -1,0 +1,1 @@
+export const cities: string[] = ['Paris', 'Cologne', 'Brussels', 'Amsterdam', 'Hamburg', 'Dusseldorf'];
