+++ project/src/consts/city-card-type.ts
@@ -1,0 +1,5 @@
+export const cityCardType: {[type: string]: string} = {
+  CITIES_CARD: 'cities__card',
+  FAVORITES_CARD: 'favorites__card',
+};
+
