+++ project/src/consts/hotel-type.ts
@@ -1,0 +1,6 @@
+export const hotelType: {[type: string]: string} = {
+  apartment: 'Apartment',
+  room: 'Private Room',
+  house: 'House',
+  hotel: 'Hotel',
+};
