+++ project/src/consts/sort-type.ts
@@ -1,0 +1,6 @@
+export const sortType: {[type: string]: string} = {
+  POPULAR: 'Popular',
+  LOW_TO_HIGH: 'Price: low to high',
+  HIGH_TO_LOW: 'Price: high to low',
+  TOP_RATED: 'Top rated first',
+};
