+++ project/src/mock/hotels.ts
@@ -5590,9 +5590,3 @@ const hotels: Hotel[] =
 ];
 
 export const getHotels = () => hotels;
-
-export const getHotelsByCity = (city: string) => hotels.filter((item: Hotel) => item.city.name === city);
-
-export const getHotelById = (id: number) => hotels.find((item: Hotel) => item.id === id);
-
-export const getFavoriteHotels = () => hotels.filter((item: Hotel) => item.isFavorite);
